Move cursor/bulk-write telemetry to network.py, centralising _CommandTelemetry
bulk_write_command and cursor_operation become free functions in network.py
(alongside the existing command function), consolidating all _CommandTelemetry
usage in one module. _CURSOR_DOC_FIELDS moves there too. AsyncConnection loses
those methods and the imports they required (_CommandTelemetry, timedelta,
Callable, _GetMore, _Query, _DocumentOut).
